pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "setlist-maker"
version = "1.0.0"
description = "Generate tracklists from DJ sets or long audio recordings using Shazam"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ray"}
]
keywords = ["dj", "tracklist", "shazam", "audio", "music", "setlist"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
]
dependencies = [
    "shazamio>=0.6.0",
    "pydub>=0.25.0",
    "audioop-lts>=0.2;python_version>='3.13'",
    "textual>=0.47.0",
    "mutagen>=1.47.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
setlist-maker = "setlist_maker.cli:main"

[project.urls]
Homepage = "https://github.com/brigleb/setlist-maker"
Issues = "https://github.com/brigleb/setlist-maker/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
